Split commands into new files
