+++ .cursorrules
@@ -1,0 +1,14 @@
+- We're building an extension for the Zed text editor:
+  - https://zed.dev/docs/extensions/developing-extensions
+  - https://github.com/zed-industries/zed/tree/main/extensions/slash-commands-example
+- We're using Rust:
+  - https://doc.rust-lang.org/stable/
+- The extension is based off of a similar VS Code extension:
+  - https://github.com/StarlaneStudios/vscode-comment-anchors
+  - https://marketplace.visualstudio.com/items?itemName=ExodiusStudios.comment-anchors
+- We'll try to implement as many features from the original as we can with Zed's extension API:
+  - https://docs.rs/zed_extension_api/latest/zed_extension_api/index.html
+- Do not use any Rust dependencies that are not already defined in the project's `Cargo.toml`.
+  - Feel free to suggest new dependencies only if they are well-maintained and significantly helpful/useful for this specific project.
+- Format all files after any changes are made using `cargo fmt`
+- Make sure the project builds after any changes using `cargo build`
