+++ src/lib.rs
@@ -1,7 +1,7 @@
 // https://zed.dev/docs/extensions/developing-extensions#webassembly
-use zed_extension_api::{
-    self as zed, SlashCommand, SlashCommandOutput, SlashCommandOutputSection, Worktree,
-};
+use zed_extension_api::{self as zed, SlashCommand, SlashCommandOutput, Worktree};
+
+mod commands;
 
 struct CommentAnchors;
 
@@ -14,48 +14,9 @@ impl zed::Extension for CommentAnchors {
         &self,
         command: SlashCommand,
         args: Vec<String>,
-        _worktree: Option<&Worktree>,
+        worktree: Option<&Worktree>,
     ) -> Result<SlashCommandOutput, String> {
-        match command.name.as_str() {
-            "echo" => {
-                if args.is_empty() {
-                    return Err("nothing to echo".to_string());
-                }
-
-                let text = args.join(" ");
-
-                Ok(SlashCommandOutput {
-                    sections: vec![SlashCommandOutputSection {
-                        range: (0..text.len()).into(),
-                        label: "Echo".to_string(),
-                    }],
-                    text,
-                })
-            }
-            "pick-one" => {
-                let Some(selection) = args.first() else {
-                    return Err("no option selected".to_string());
-                };
-
-                match selection.as_str() {
-                    "option-1" | "option-2" | "option-3" => {}
-                    invalid_option => {
-                        return Err(format!("{invalid_option} is not a valid option"));
-                    }
-                }
-
-                let text = format!("You chose {selection}.");
-
-                Ok(SlashCommandOutput {
-                    sections: vec![SlashCommandOutputSection {
-                        range: (0..text.len()).into(),
-                        label: format!("Pick One: {selection}"),
-                    }],
-                    text,
-                })
-            }
-            command => Err(format!("unknown slash command: \"{command}\"")),
-        }
+        commands::handle_command(command, args, worktree)
     }
 }
